texk/dvipdfm-x: Minor bug fix and cleanup
git-svn-id: svn://tug.org/texlive/trunk/Build/source@35246 c570f23f-e606-0410-a88d-b1316a301751

+++ texk/dvipdfm-x/bmpimage.h
@@ -29,7 +29,7 @@
 
 extern int bmp_include_image (pdf_ximage *ximage, FILE *file);
 extern int check_for_bmp     (FILE *file);
-extern int bmp_get_bbox      (FILE *file, long *width, long *height,
+extern int bmp_get_bbox      (FILE *file, int *width, int *height,
                               double *xdensity, double *ydensity);
 
 #endif /* _BMPIMAGE_H_ */
